# docs/swagger.yaml
basePath: /api
definitions:
  common.Page:
    properties:
      current:
        type: integer
      list: {}
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  define.AddRolePermissionRequest:
    properties:
      permissionIds:
        items:
          type: integer
        type: array
      roleId:
        type: integer
    required:
    - permissionIds
    - roleId
    type: object
  define.AddUserRoleRequest:
    properties:
      roleIds:
        items:
          type: integer
        type: array
      userId:
        type: integer
    required:
    - roleIds
    - userId
    type: object
...
  define.CreateRoleRequest:
    properties:
      description:
        type: string
      roleCode:
        type: string
      roleName:
        type: string
    required:
    - roleCode
    - roleName
    type: object
...
  define.DeleteUserRoleRequest:
    properties:
      roleIds:
        items:
          type: integer
        type: array
      userId:
        type: integer
    required:
    - roleIds
    - userId
    type: object
...
  define.ListUserRoleRequest:
    properties:
      userId:
        type: integer
    required:
    - userId
    type: object
  define.ListUserRoleResponse:
    properties:
      roleId:
        type: integer
      roleName:
        type: string
    type: object
...
  define.RolePageResponse:
    properties:
      description:
        type: string
      roleCode:
        type: string
      roleId:
        type: integer
      roleName:
        type: string
    type: object
...
  define.UpdateRoleRequest:
    properties:
      description:
        type: string
      roleCode:
        type: string
      roleId:
        type: integer
      roleName:
        type: string
    required:
    - roleId
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a ether api server.
  title: Ether API
  version: "1"
paths:
...
  /role:
    get:
      consumes:
      - application/json
      description: List roles
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page_size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/define.RolePageResponse'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: List roles
      tags:
      - role
    post:
      consumes:
      - application/json
      description: Create role
      parameters:
      - description: CreateRoleRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Create role
      tags:
      - role
  /role/{role_id}:
    delete:
      consumes:
      - application/json
      description: Delete role
      parameters:
      - description: role_id
        in: path
        name: role_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Delete role
      tags:
      - role
    put:
      consumes:
      - application/json
      description: Update role
      parameters:
      - description: UpdateRoleRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.UpdateRoleRequest'
      - description: role_id
        in: path
        name: role_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Update role
      tags:
      - role
  /roles/{roleId}/permissions:
    get:
      consumes:
      - application/json
      description: List role permission
      parameters:
      - description: role_id
        in: path
        name: role_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - Bearer: []
      summary: List role permission
      tags:
      - role
  /roles/{roleId}/permissions/add:
    post:
      consumes:
      - application/json
      description: Add role permission
      parameters:
      - description: role_id
        in: path
        name: role_id
        required: true
        type: integer
      - description: AddRolePermissionRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.AddRolePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Add role permission
      tags:
      - role
  /roles/{roleId}/permissions/delete:
    post:
      consumes:
      - application/json
      description: Delete role permission
      parameters:
      - description: role_id
        in: path
        name: role_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: Delete role permission
      tags:
      - role
...
  /users/{userId}/roles:
    get:
      consumes:
      - application/json
      description: List user role
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      - description: ListUserRoleRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.ListUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/define.ListUserRoleResponse'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: List user role
      tags:
      - user
  /users/{userId}/roles/add:
    post:
      consumes:
      - application/json
      description: Add user role
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      - description: AddUserRoleRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.AddUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - Bearer: []
      summary: Add user role
      tags:
      - user
  /users/{userId}/roles/delete:
    post:
      consumes:
      - application/json
      description: Delete user role
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      - description: DeleteUserRoleRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/define.DeleteUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete user role
      tags:
      - user
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"